services/enrollment/requirements.txt
pillow==10.4.0
confluent-kafka==2.5.3
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.27.2